feat: Swap order of Secret URL
feat: Swapping the order of the generated Secret URL.

This means that the complete URL will be shown first along with the QR
Code.

The Split URL is now below and hidden under a expandable menu.

This is done to ensure users don't mistakenly send the URL without
encryption key.