feat: Hide file upload when not signed in
feat: Hiding the File Upload section when a user is not signed in.

This change is made as even if you disable the file upload in settings
it will still show a section mentioning to sign in to upload files.

Removing this so that if the option is disabled the section is hidden
entirely.