✨ feat: Add var to enable advanced copy
In most cases, users won't be copying the exposed password in HTML or
base64, and as such these options are not really needed.
As a way around this, we're enabling a environment variable to control
this - `SECRET_ADVANCED_COPY` - this value is set to `false` by default
which means there will be a simple "Copy" button only. IF in the future
base64 or HTML is required, this can be re-enabled by the env var.