fix: API URL Mismatch
We're facing an issue where when we call the API's the URL that is
returned is based off the `HOST_URL` environment variable rather than
the host we're curling.

This leads to a URL mismatch as we curl 1 endpoint, but the response is
another.

This fix will ignore the HOST_URL variable in this case, and simply
read the URL you curl and pass that on.